separated index page into a few components
