+++ src/components/Navbar.tsx
@@ -10,7 +10,6 @@ const Nav = styled.nav`
     justify-content: space-between;
     align-items: center;
     flex-wrap: wrap;
-    height: 98px;
     background-color: ${props => props.theme.colors.navBackground};
     color: ${props => props.theme.colors.text};
     z-index: 1000;
